requirements.txt
google-api-python-client==2.136.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib==1.2.0
python-dotenv==0.21.1
pandas>=1.3.0
openpyxl>=3.0.7
python-dateutil>=2.8.2
tqdm==4.66.4 